@@ -1,20 +1,30 @@
 <?php
 
-use Sylius\Bundle\CoreBundle\Kernel\Kernel;
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
+use Sylius\Bundle\CoreBundle\Application\Kernel;
+
+/**
+ * @author Paweł Jędrzejewski <pawel@sylius.org>
+ * @author Gonzalo Vilaseca <gvilaseca@reiss.co.uk>
+ */
 class AppKernel extends Kernel
 {
+    /**
+     * {@inheritdoc}
+     */
     public function registerBundles()
     {
-        $bundles = array(
-            new \Sensio\Bundle\GeneratorBundle\SensioGeneratorBundle(),
-            new \AppBundle\AppBundle()
-        );
+        $bundles = [
 
-        if (in_array($this->environment, array('dev', 'test'))) {
-            $bundles[] = new Symfony\Bundle\WebProfilerBundle\WebProfilerBundle();
-            $bundles[] = new Symfony\Bundle\DebugBundle\DebugBundle();
-        }
+        ];
 
         return array_merge(parent::registerBundles(), $bundles);
     }
