@@ -1,6 +1,13 @@
 <?php
 
-require_once __DIR__.'/AppKernel.php';
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
 use Symfony\Bundle\FrameworkBundle\HttpCache\HttpCache;
 
