@@ -1,22 +1,28 @@
 <?php
 
-use Symfony\Component\ClassLoader\ApcClassLoader;
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 use Symfony\Component\HttpFoundation\Request;
 
-$loader = require_once __DIR__.'/../app/bootstrap.php.cache';
+/*
+ * Sylius front controller.
+ * Live (production) environment.
+ */
 
-//$loader = new ApcClassLoader('sylius', $loader);
-//$loader->register(true);
-
+require_once __DIR__.'/../app/bootstrap.php.cache';
 require_once __DIR__.'/../app/AppKernel.php';
-//require_once __DIR__.'/../app/AppCache.php';
 
 $kernel = new AppKernel('prod', false);
 $kernel->loadClassCache();
-//$kernel = new AppCache($kernel);
 
 $request = Request::createFromGlobals();
-Request::enableHttpMethodParameterOverride();
 
 $response = $kernel->handle($request);
 $response->send();
