@@ -1,11 +1,31 @@
 <?php
 
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+use Symfony\Component\Debug\Debug;
 use Symfony\Component\HttpFoundation\Request;
 
-if (isset($_SERVER['HTTP_CLIENT_IP'])
+/*
+ * Sylius front controller.
+ * Dev environment.
+ *
+ * To develop on Sylius in Vagrant set the SYLIUS_APP_DEV_PERMITTED to a non zero value.
+ * e.g. in apache, through your vhost configuration file:
+ *
+ *   SetEnv SYLIUS_APP_DEV_PERMITTED 1
+ */
+if (!getenv("SYLIUS_APP_DEV_PERMITTED") && (
+    isset($_SERVER['HTTP_CLIENT_IP'])
     || isset($_SERVER['HTTP_X_FORWARDED_FOR'])
-    || !(in_array(@$_SERVER['REMOTE_ADDR'], array('127.0.0.1', 'fe80::1', '::1')) || php_sapi_name() === 'cli-server')
-) {
+    || !(in_array(@$_SERVER['REMOTE_ADDR'], ['127.0.0.1', 'fe80::1', '::1', '113.0.0.1'], true) || php_sapi_name() === 'cli-server')
+)) {
     header('HTTP/1.0 403 Forbidden');
     exit('You are not allowed to access this file. Check '.basename(__FILE__).' for more information.');
 }
@@ -13,11 +33,11 @@ if (isset($_SERVER['HTTP_CLIENT_IP'])
 require_once __DIR__.'/../app/bootstrap.php.cache';
 require_once __DIR__.'/../app/AppKernel.php';
 
+Debug::enable();
+
 $kernel = new AppKernel('dev', true);
-$kernel->loadClassCache();
 
 $request = Request::createFromGlobals();
-Request::enableHttpMethodParameterOverride();
 
 $response = $kernel->handle($request);
 $response->send();
