@@ -9,12 +9,11 @@
  * file that was distributed with this source code.
  */
 
-use Symfony\Component\Debug\Debug;
 use Symfony\Component\HttpFoundation\Request;
 
 /*
  * Sylius front controller.
- * Testing dev-like environment.
+ * Testing prod-like environment.
  */
 
 if (isset($_SERVER['HTTP_CLIENT_IP'])
@@ -28,9 +27,8 @@ if (isset($_SERVER['HTTP_CLIENT_IP'])
 require_once __DIR__.'/../app/bootstrap.php.cache';
 require_once __DIR__.'/../app/AppKernel.php';
 
-Debug::enable();
-
-$kernel = new AppKernel('test', true);
+$kernel = new AppKernel('test_cached', false);
+$kernel->loadClassCache();
 
 $request = Request::createFromGlobals();
 
